requirements.txt
curl_cffi
pyyaml
